+++ idea/testData/intentions/introduceVariable/atExpressionEnd.kt
@@ -1,0 +1,5 @@
+fun foo(a: Int, b: Int) = a + b
+
+fun foo() {
+    foo(1 + 2, 3 * 4)<caret>
+}