+++ idea/testData/intentions/introduceVariable/nothingType.kt
@@ -1,0 +1,4 @@
+// IS_APPLICABLE: false
+fun foo(): Int {
+   return 1 + 2<caret>
+}