+++ idea/testData/intentions/introduceVariable/unitType.kt
@@ -1,0 +1,6 @@
+// IS_APPLICABLE: false
+fun foo(a: Int, b: Int) {}
+
+fun foo() {
+   foo(1 + 2, 3 * 4)<caret>
+}